# Improve Fortunes bench performance for Servant