--- a/frameworks/Haskell/servant/mysql-haskell/stack.yaml
+++ b/frameworks/Haskell/servant/mysql-haskell/stack.yaml
@@ -1,7 +1,0 @@
-resolver: lts-13.11
-packages:
-- '.'
-
-# the following flags are meant for use with ../servant-mysql-haskell.dockerfile
-compiler: ghc-8.6.3 # this MUST match the resolver's GHC; minor hack to ensure GHC isn't downloaded into sandbox.
-allow-different-user: true # carryover from hasql sibling test dir