--- a/frameworks/Haskell/servant/stack.yaml
+++ b/frameworks/Haskell/servant/stack.yaml
@@ -1,7 +1,8 @@
-resolver: lts-13.10
+resolver: lts-13.13
 packages:
-- '.'
+- './hasql'
+- './mysql-haskell'
 
-# the following flags are meant for use with ../servant-mysql-haskell.dockerfile
+# the following flags are meant for use with servant.dockerfile
 compiler: ghc-8.6.3 # this MUST match the resolver's GHC; minor hack to ensure GHC isn't downloaded into sandbox.
-allow-different-user: true # carryover from hasql sibling test dir
+allow-different-user: true