- Switched html rendering to [type-of-html](https://hackage.haskell.org/package/type-of-html) to boost Fortunes test performance.
- Travis run ~in progress~ has timed out: https://travis-ci.org/TechEmpower/FrameworkBenchmarks/jobs/507477716
- Performance is better (than [ffec2c1](https://github.com/TechEmpower/FrameworkBenchmarks/pull/4550/commits/ffec2c1c25f34deebb4017a438b854cb5852329d)) for Fortunes test by about 8%.
- Results from running them on my laptop. Note: had to rename `servant-mysql-haskell` strings in results.json to `servant` to get the visualization UI to render results. Attached results has this rename in place for easier visualization.
  - <summary>`servant-mysql-haskell` from `ffec2c1` 
    <details>

      ```json
      {
        "rawData": {
          "fortune": {
            "servant": [
              {
                "latencyAvg": "3.68ms", 
                "timeout": 15, 
                "latencyMax": "33.37ms", 
                "latencyStdev": "2.59ms", 
                "totalRequests": 69771, 
                "write": 0, 
                "read": 0, 
                "connect": 0, 
                "startTime": 1552441115, 
                "endTime": 1552441130
              }, 
              {
                "latencyAvg": "6.70ms", 
                "latencyMax": "55.29ms", 
                "latencyStdev": "3.55ms", 
                "totalRequests": 72269, 
                "startTime": 1552441132, 
                "endTime": 1552441147
              }, 
              {
                "latencyAvg": "14.34ms", 
                "timeout": 47, 
                "latencyMax": "131.48ms", 
                "latencyStdev": "9.86ms", 
                "totalRequests": 71155, 
                "write": 0, 
                "read": 0, 
                "connect": 0, 
                "startTime": 1552441149, 
                "endTime": 1552441164
              }, 
              {
                "latencyAvg": "29.00ms", 
                "latencyMax": "302.46ms", 
                "latencyStdev": "19.78ms", 
                "totalRequests": 70795, 
                "startTime": 1552441166, 
                "endTime": 1552441182
              }, 
              {
                "latencyAvg": "54.84ms", 
                "timeout": 72, 
                "latencyMax": "492.49ms", 
                "latencyStdev": "37.10ms", 
                "totalRequests": 72704, 
                "write": 0, 
                "read": 0, 
                "connect": 0, 
                "startTime": 1552441184, 
                "endTime": 1552441199
              }, 
              {
                "latencyAvg": "115.98ms", 
                "latencyMax": "1.16s", 
                "latencyStdev": "55.56ms", 
                "totalRequests": 67307, 
                "startTime": 1552441201, 
                "endTime": 1552441216
              }
            ]
          }, 
          "plaintext": {
            "servant": [
              {
                "latencyAvg": "139.12ms", 
                "latencyMax": "708.06ms", 
                "latencyStdev": "100.49ms", 
                "totalRequests": 333172, 
                "startTime": 1552441250, 
                "endTime": 1552441265
              }, 
              {
                "latencyAvg": "393.27ms", 
                "latencyMax": "2.77s", 
                "latencyStdev": "234.77ms", 
                "totalRequests": 350484, 
                "startTime": 1552441267, 
                "endTime": 1552441283
              }, 
              {
                "latencyAvg": "2.34s", 
                "timeout": 55, 
                "latencyMax": "7.78s", 
                "latencyStdev": "1.55s", 
                "totalRequests": 251841, 
                "write": 0, 
                "read": 0, 
                "connect": 0, 
                "startTime": 1552441285, 
                "endTime": 1552441300
              }, 
              {
                "latencyAvg": "4.57s", 
                "timeout": 1859, 
                "latencyMax": "7.70s", 
                "latencyStdev": "2.05s", 
                "totalRequests": 174522, 
                "write": 0, 
                "read": 0, 
                "connect": 0, 
                "startTime": 1552441302, 
                "endTime": 1552441317
              }
            ]
          }, 
          "slocCounts": {
            "servant": 1099
          }, 
          "db": {
            "servant": [
              {
                "latencyAvg": "3.29ms", 
                "timeout": 16, 
                "latencyMax": "66.44ms", 
                "latencyStdev": "2.66ms", 
                "totalRequests": 78119, 
                "write": 0, 
                "read": 0, 
                "connect": 0, 
                "startTime": 1552440609, 
                "endTime": 1552440624
              }, 
              {
                "latencyAvg": "6.35ms", 
                "latencyMax": "216.55ms", 
                "latencyStdev": "5.41ms", 
                "totalRequests": 78942, 
                "startTime": 1552440626, 
                "endTime": 1552440641
              }, 
              {
                "latencyAvg": "12.57ms", 
                "timeout": 64, 
                "latencyMax": "105.59ms", 
                "latencyStdev": "6.02ms", 
                "totalRequests": 78138, 
                "write": 0, 
                "read": 0, 
                "connect": 0, 
                "startTime": 1552440643, 
                "endTime": 1552440658
              }, 
              {
                "latencyAvg": "25.01ms", 
                "latencyMax": "293.58ms", 
                "latencyStdev": "12.44ms", 
                "totalRequests": 78711, 
                "startTime": 1552440660, 
                "endTime": 1552440675
              }, 
              {
                "latencyAvg": "50.96ms", 
                "timeout": 31, 
                "latencyMax": "1.10s", 
                "latencyStdev": "37.50ms", 
                "totalRequests": 78609, 
                "write": 0, 
                "read": 0, 
                "connect": 0, 
                "startTime": 1552440677, 
                "endTime": 1552440692
              }, 
              {
                "latencyAvg": "137.10ms", 
                "latencyMax": "2.31s", 
                "latencyStdev": "200.75ms", 
                "totalRequests": 74615, 
                "startTime": 1552440694, 
                "endTime": 1552440710
              }
            ]
          }, 
          "update": {
            "servant": [
              {
                "latencyAvg": "550.82ms", 
                "latencyMax": "1.67s", 
                "latencyStdev": "219.87ms", 
                "totalRequests": 13562, 
                "startTime": 1552440997, 
                "endTime": 1552441012
              }, 
              {
                "latencyAvg": "2.09s", 
                "latencyMax": "4.49s", 
                "latencyStdev": "416.70ms", 
                "totalRequests": 3380, 
                "startTime": 1552441014, 
                "endTime": 1552441030
              }, 
              {
                "latencyAvg": "2.66s", 
                "latencyMax": "6.74s", 
                "latencyStdev": "1.04s", 
                "totalRequests": 2616, 
                "startTime": 1552441032, 
                "endTime": 1552441047
              }, 
              {
                "latencyAvg": "4.22s", 
                "latencyMax": "7.05s", 
                "latencyStdev": "868.28ms", 
                "totalRequests": 1533, 
                "startTime": 1552441049, 
                "endTime": 1552441064
              }, 
              {
                "latencyAvg": "4.70s", 
                "timeout": 6, 
                "latencyMax": "7.95s", 
                "latencyStdev": "1.19s", 
                "totalRequests": 1355, 
                "write": 0, 
                "read": 0, 
                "connect": 0, 
                "startTime": 1552441066, 
                "endTime": 1552441081
              }
            ]
          }, 
          "json": {
            "servant": [
              {
                "latencyAvg": "1.30ms", 
                "latencyMax": "24.83ms", 
                "latencyStdev": "1.23ms", 
                "totalRequests": 196889, 
                "startTime": 1552440744, 
                "endTime": 1552440759
              }, 
              {
                "latencyAvg": "2.89ms", 
                "timeout": 28, 
                "latencyMax": "44.37ms", 
                "latencyStdev": "4.01ms", 
                "totalRequests": 249075, 
                "write": 0, 
                "read": 0, 
                "connect": 0, 
                "startTime": 1552440761, 
                "endTime": 1552440776
              }, 
              {
                "latencyAvg": "4.58ms", 
                "latencyMax": "44.69ms", 
                "latencyStdev": "5.61ms", 
                "totalRequests": 303806, 
                "startTime": 1552440778, 
                "endTime": 1552440793
              }, 
              {
                "latencyAvg": "7.37ms", 
                "timeout": 128, 
                "latencyMax": "73.26ms", 
                "latencyStdev": "6.96ms", 
                "totalRequests": 293819, 
                "write": 0, 
                "read": 0, 
                "connect": 0, 
                "startTime": 1552440795, 
                "endTime": 1552440810
              }, 
              {
                "latencyAvg": "12.74ms", 
                "latencyMax": "71.46ms", 
                "latencyStdev": "7.69ms", 
                "totalRequests": 292854, 
                "startTime": 1552440812, 
                "endTime": 1552440828
              }, 
              {
                "latencyAvg": "24.71ms", 
                "latencyMax": "114.99ms", 
                "latencyStdev": "9.59ms", 
                "totalRequests": 295910, 
                "startTime": 1552440830, 
                "endTime": 1552440845
              }
            ]
          }, 
          "commitCounts": {
            "servant": 0
          }, 
          "query": {
            "servant": [
              {
                "latencyAvg": "158.65ms", 
                "latencyMax": "2.90s", 
                "latencyStdev": "124.92ms", 
                "totalRequests": 51712, 
                "startTime": 1552440879, 
                "endTime": 1552440894
              }, 
              {
                "latencyAvg": "428.62ms", 
                "latencyMax": "2.84s", 
                "latencyStdev": "126.83ms", 
                "totalRequests": 17592, 
                "startTime": 1552440896, 
                "endTime": 1552440911
              }, 
              {
                "latencyAvg": "695.50ms", 
                "latencyMax": "3.04s", 
                "latencyStdev": "137.90ms", 
                "totalRequests": 10716, 
                "startTime": 1552440913, 
                "endTime": 1552440929
              }, 
              {
                "latencyAvg": "1.02s", 
                "latencyMax": "4.73s", 
                "latencyStdev": "247.18ms", 
                "totalRequests": 7307, 
                "startTime": 1552440931, 
                "endTime": 1552440946
              }, 
              {
                "latencyAvg": "1.30s", 
                "latencyMax": "3.11s", 
                "latencyStdev": "189.31ms", 
                "totalRequests": 5625, 
                "startTime": 1552440948, 
                "endTime": 1552440963
              }
            ]
          }, 
          "cached_query": {}
        }, 
        "environmentDescription": "(unspecified, hostname = ec65b9a95fbf)", 
        "git": {
          "commitId": "ffec2c1c25f34deebb4017a438b854cb5852329d", 
          "repositoryUrl": "git@github.com:naushadh/FrameworkBenchmarks.git", 
          "branchName": "servant/mysql-haskell"
        }, 
        "cachedQueryIntervals": [
          1, 
          10, 
          20, 
          50, 
          100
        ], 
        "uuid": "c3a79cb8-d252-4e69-b59c-b966389f1164", 
        "succeeded": {
          "fortune": [
            "servant"
          ], 
          "plaintext": [
            "servant"
          ], 
          "db": [
            "servant"
          ], 
          "update": [
            "servant"
          ], 
          "json": [
            "servant"
          ], 
          "query": [
            "servant"
          ], 
          "cached_query": []
        }, 
        "failed": {
          "fortune": [], 
          "plaintext": [], 
          "db": [], 
          "update": [], 
          "json": [], 
          "query": [], 
          "cached_query": []
        }, 
        "verify": {
          "servant": {
            "fortune": "pass", 
            "plaintext": "pass", 
            "db": "pass", 
            "update": "pass", 
            "json": "pass", 
            "query": "pass"
          }
        }, 
        "duration": 15, 
        "frameworks": [
          "servant"
        ], 
        "pipelineConcurrencyLevels": [
          256, 
          1024, 
          4096, 
          16384
        ], 
        "completionTime": 1552441330047, 
        "concurrencyLevels": [
          16, 
          32, 
          64, 
          128, 
          256, 
          512
        ], 
        "startTime": 1552440493660, 
        "queryIntervals": [
          1, 
          5, 
          10, 
          15, 
          20
        ], 
        "completed": {
          "servant": "20190313014205"
        }, 
        "name": "(unspecified, datetime = 2019-03-13 01:28:13)"
      }
      ```

    </details>
    </summary>
  - <summary>`servant-mysql-haskell` from `b74090b` 
    <details>

      ```json
      {
        "rawData": {
          "fortune": {
            "servant": [
              {
                "latencyAvg": "3.42ms", 
                "latencyMax": "76.15ms", 
                "latencyStdev": "2.76ms", 
                "totalRequests": 75939, 
                "startTime": 1552792222, 
                "endTime": 1552792237
              }, 
              {
                "latencyAvg": "6.24ms", 
                "timeout": 32, 
                "latencyMax": "64.95ms", 
                "latencyStdev": "3.67ms", 
                "totalRequests": 78603, 
                "write": 0, 
                "read": 0, 
                "connect": 0, 
                "startTime": 1552792239, 
                "endTime": 1552792255
              }, 
              {
                "latencyAvg": "13.67ms", 
                "latencyMax": "148.41ms", 
                "latencyStdev": "8.98ms", 
                "totalRequests": 74397, 
                "startTime": 1552792257, 
                "endTime": 1552792272
              }, 
              {
                "latencyAvg": "25.30ms", 
                "latencyMax": "265.06ms", 
                "latencyStdev": "14.27ms", 
                "totalRequests": 77420, 
                "startTime": 1552792274, 
                "endTime": 1552792289
              }, 
              {
                "latencyAvg": "53.14ms", 
                "timeout": 28, 
                "latencyMax": "267.95ms", 
                "latencyStdev": "31.38ms", 
                "totalRequests": 73622, 
                "write": 0, 
                "read": 0, 
                "connect": 0, 
                "startTime": 1552792291, 
                "endTime": 1552792306
              }, 
              {
                "latencyAvg": "111.32ms", 
                "latencyMax": "1.36s", 
                "latencyStdev": "64.59ms", 
                "totalRequests": 71187, 
                "startTime": 1552792308, 
                "endTime": 1552792323
              }
            ]
          }, 
          "plaintext": {
            "servant": [
              {
                "latencyAvg": "116.86ms", 
                "latencyMax": "327.78ms", 
                "latencyStdev": "60.85ms", 
                "totalRequests": 290158, 
                "startTime": 1552792357, 
                "endTime": 1552792372
              }, 
              {
                "latencyAvg": "488.30ms", 
                "latencyMax": "4.91s", 
                "latencyStdev": "394.55ms", 
                "totalRequests": 326344, 
                "startTime": 1552792374, 
                "endTime": 1552792390
              }, 
              {
                "latencyAvg": "2.11s", 
                "timeout": 48, 
                "latencyMax": "7.66s", 
                "latencyStdev": "1.34s", 
                "totalRequests": 270574, 
                "write": 0, 
                "read": 0, 
                "connect": 0, 
                "startTime": 1552792392, 
                "endTime": 1552792407
              }, 
              {
                "latencyAvg": "3.80s", 
                "timeout": 3259, 
                "latencyMax": "7.48s", 
                "latencyStdev": "1.68s", 
                "totalRequests": 193791, 
                "write": 0, 
                "read": 0, 
                "connect": 0, 
                "startTime": 1552792409, 
                "endTime": 1552792424
              }
            ]
          }, 
          "slocCounts": {
            "servant": 710
          }, 
          "db": {
            "servant": [
              {
                "latencyAvg": "3.51ms", 
                "latencyMax": "58.91ms", 
                "latencyStdev": "3.01ms", 
                "totalRequests": 75671, 
                "startTime": 1552791715, 
                "endTime": 1552791730
              }, 
              {
                "latencyAvg": "7.16ms", 
                "timeout": 31, 
                "latencyMax": "95.72ms", 
                "latencyStdev": "4.79ms", 
                "totalRequests": 70111, 
                "write": 0, 
                "read": 0, 
                "connect": 0, 
                "startTime": 1552791732, 
                "endTime": 1552791747
              }, 
              {
                "latencyAvg": "11.20ms", 
                "timeout": 61, 
                "latencyMax": "137.90ms", 
                "latencyStdev": "6.02ms", 
                "totalRequests": 87958, 
                "write": 0, 
                "read": 0, 
                "connect": 0, 
                "startTime": 1552791749, 
                "endTime": 1552791765
              }, 
              {
                "latencyAvg": "22.98ms", 
                "latencyMax": "249.93ms", 
                "latencyStdev": "8.06ms", 
                "totalRequests": 84148, 
                "startTime": 1552791767, 
                "endTime": 1552791782
              }, 
              {
                "latencyAvg": "52.96ms", 
                "timeout": 2, 
                "latencyMax": "402.92ms", 
                "latencyStdev": "20.05ms", 
                "totalRequests": 72553, 
                "write": 0, 
                "read": 0, 
                "connect": 0, 
                "startTime": 1552791784, 
                "endTime": 1552791799
              }, 
              {
                "latencyAvg": "110.67ms", 
                "latencyMax": "1.68s", 
                "latencyStdev": "91.84ms", 
                "totalRequests": 74888, 
…