@naushadh 2019-03-17
Branched off of the same point as https://github.com/TechEmpower/FrameworkBenchmarks/pull/4550. The diff should make more sense after the parent PR is merged.