@nbrady-techempower 2019-03-20
@naushadh The servant build time is exceeding the 50 min time limit imposed by travis. Is it possible to speed this build up?