@naushadh 2019-03-21
I've noticed that. Each test within `servant` finishes in ~25 minutes. it's only when all tests are run under servant as a whole we max out 50 minutes. Is there a way we can alter the Travis matrix to run each test within servant separately?