@nbrady-techempower 2019-03-21
We don't necessarily want to do that. We use that 50 mins as limit that frameworks should be under so they don't hoard time when running in our continuous environments. Is there anything you can do to combine them such that the first few build steps for each dockerfile are the same that way the docker cache is utilized? Like copying the `stack.yml` from a shared folder and doing `RUN stack setup` from that? Maybe the first several lines could look something like:

```
FROM haskell:8.6.3

RUN apt update -yqq && apt install -yqq xz-utils make libpq-dev

WORKDIR /servant

COPY ./shared/stack.yml

RUN stack setup

# Then from here on you'd do something different for each framework like

COPY ./hasql/servant-bench.cabal .
RUN stack install --dependencies-only
```

I don't know much about Haskell, but if you can run `stack setup` before copying the cabal file this might cut your time down.