@nbrady-techempower 2019-03-22
@naushadh additionally, we're adding a check in the toolset #4569 to make sure individual builds don't exceed 60 minutes. This will probably be fine there, since it's separate tests. Still, would like to get this passing travis under 50 minutes.