@naushadh 2019-03-22
I see. I know how to speed it up to get it under 50 (actually under 30) minutes.