@naushadh 2019-03-24
@nbrady-techempower servant now completes in ~34 minutes:
https://travis-ci.org/TechEmpower/FrameworkBenchmarks/jobs/510563406