@naushadh 2019-04-09
Comparing last run [results](https://github.com/TechEmpower/FrameworkBenchmarks/pull/4550#issuecomment-475818264) based on 96c5736 against [next run](https://tfb-status.techempower.com/results/8ab47a8b-0b0d-43f7-ad12-1e96b72ff42d) based on on dc82e00. 

.|JSON Serialization|Single Query|Mutiple queries|Fortunes|Updates|Plaintext
---|---|---|---|---|---|---
servant (96c5736)|297,444|115,204|14,958|91,456|2,584|340,137
servant (dc82e00)|296,528|115,289|15,557|100,514|2,577|337,742
servant-mysql-haskell (96c5736)|301,422|155,806|10,260|97,185|2,572|344,312
servant-mysql-haskell (dc82e00)|300,650|157,379|10,146|108,146|2,611|342,074